change the cost based index planner it: change the metadata file to spark-it project; delete cube_planner test data
