--- src/examples/test_case_data/localmeta/metadata/_global/project/cube_planner.json
@@ -1,35 +1,0 @@
-{
-  "uuid" : "17bfd5ad-76ee-3d09-ebe1-882ba7598f27",
-  "last_modified" : 1675840447178,
-  "create_time" : 1675840447174,
-  "version" : "4.0.0.0",
-  "name" : "cube_planner",
-  "owner" : "ADMIN",
-  "status" : "ENABLED",
-  "create_time_utc" : 1675840447175,
-  "default_database" : "DEFAULT",
-  "description" : "",
-  "principal" : null,
-  "keytab" : null,
-  "maintain_model_type" : "MANUAL_MAINTAIN",
-  "override_kylin_properties" : {
-    "kylin.metadata.semi-automatic-mode" : "false",
-    "kylin.query.metadata.expose-computed-column" : "true",
-    "kylin.source.default" : "9"
-  },
-  "segment_config" : {
-    "auto_merge_enabled" : false,
-    "auto_merge_time_ranges" : [ "WEEK", "MONTH", "QUARTER", "YEAR" ],
-    "volatile_range" : {
-      "volatile_range_number" : 0,
-      "volatile_range_enabled" : false,
-      "volatile_range_type" : "DAY"
-    },
-    "retention_range" : {
-      "retention_range_number" : 1,
-      "retention_range_enabled" : false,
-      "retention_range_type" : "MONTH"
-    },
-    "create_empty_segment_enabled" : false
-  }
-}
