+++ src/query/src/test/java/org/apache/kylin/query/engine/AsyncQueryJobTest.java
@@ -217,7 +217,7 @@ public class AsyncQueryJobTest extends NLocalFileMetadataTestCase {
                 rawResourceMap.put(zipEntry.getName(), raw);
             }
         }
-        Assert.assertEquals(85, rawResourceMap.size());
+        Assert.assertEquals(84, rawResourceMap.size());
     }
 
     private void testKylinConfig(FileSystem workingFileSystem, FileStatus metaFileStatus) throws IOException {
