+++ src/spark-project/spark-it/src/test/resources/index_planner/index_plan/d863b37c-e1a9-717f-7df7-74991815b1eb.json
